pyproject.toml
[project]
name = "djereo"
version = "3.18.0"
description = ""
authors = [
    {name = "Alberto Morón Hernández", email = "alberto.moronh@gmail.com"}
]
readme = "README.md"
license-files = ["LICENSE"]
requires-python = ">=3.13"
dependencies = [
    "ipython>=9.7.0,<10.0.0",
]

[dependency-groups]
test = [
    "beautifulsoup4>=4.14.2,<5.0.0",
    "copier>=9.11.0,<10.0.0",
    "pip>=26.0.1,<27.0.0",
    "prek>=0.3.2,<0.4.0",
    "pytest>=9.0.0,<10.0.0",
    "pytest-randomly>=4.0.1,<5.0.0",
    "pytest-xdist>=3.8.0,<4.0.0",
    "pyyaml>=6.0.3,<7.0.0",
    "psycopg[binary]>=3.3.0,<4.0.0",
    "sh>=2.2.2,<3.0.0",
]

[project.urls]
Changelog = "https://github.com/albertomh/djereo/blob/main/CHANGELOG.md"
Documentation = "https://albertomh.github.io/djereo/"
Repository = "https://github.com/albertomh/djereo"

[tool.setuptools.packages.find]
# allow `pip install -e .` to find the project package
include = ["djereo"]

[tool.pytest]
addopts = ["--pdbcls=IPython.terminal.debugger:Pdb"]
markers = [
    "integration",
    "skip_migrate",
    "slow",
    "smoke",
    "unit",
]

[tool.ruff]
exclude = [".pdbrc"]
extend-exclude = ["*project_name*"]
force-exclude = true
indent-width = 4
line-length = 90

[tool.ruff.lint]
ignore = []
select = [
    # flake8-bugbear
    "B",
    # pycodestyle errors
    "E",
    # isort
    "I",
    # flake8-no-pep420
    "INP",
    # Pyflakes
    "F",
    #flake8-logging-format
    "G",
    # flake8-logging
    "LOG",
    # pylint-pl
    "PL",
    # ruff-specific-rules-ruf
    "RUF",
    # flake8-bandit
    "S",
    # flake8-tidy-imports
    "TID",
    # tryceratops-try
    "TRY",
    # flake8-print
    "T20",
    # pyupgrade-up
    "UP",
    # pycodestyle warnings
    "W",
]

[tool.ruff.lint.isort]
known-first-party = ["users", "core"]

[tool.ruff.lint.per-file-ignores]
"noxfile.py" = ["S"]
"tests/**.py" = ["PLC0415", "PLR0913", "S"]
"**/tests/**.py" = ["S"]
"**/tests_e2e/**.py" = ["S"]
